Removed userId from the application-sites relation
