+++ packages/api/curl-scripts/README.md
@@ -1,0 +1,1 @@
+For testing
