+++ packages/api/curl-scripts/setup-sites.sh
@@ -21,3 +21,8 @@ curl -X 'POST' \
   -d '{
   "name": "site 3"
 }'
+
+curl -X 'GET' \
+  'http://localhost:3000/sites' \
+  -H 'accept: application/json' \
+  -H 'Content-Type: application/json' | node prettify.js
