+++ packages/api/src/handlers/application-site/__tests__/application-site.spec.js
@@ -4,7 +4,6 @@
 
 const APPLICATION_ID = '4f2b56c5-1d2c-4a4b-b4ad-e4a1ac90d5c8'
 const SITE_ID = '5d152a82-067b-4162-831e-c5b5e19cff7f'
-const USER_ID = '8d1adfa1-8451-4109-a71c-9f7aa516a3f1'
 const APPLICATION_SITE_ID = '4b0ccdd8-53af-414e-9f2a-a9c61493ab97'
 
 const path = 'user/8d1adfa1-8451-4109-a71c-9f7aa516a3f1/application-site'
@@ -36,7 +35,7 @@ const tsR = {
 /*
  * Create the parameters to mock the openApi context which is inserted into each handler
  */
-const context = { request: { params: { userId: USER_ID } } }
+const context = { request: { } }
 
 jest.mock('@defra/wls-database-model')
 
@@ -54,14 +53,12 @@ describe('The postApplicationSite handler', () => {
   it('returns a 201 on successful create', async () => {
     cache.save = jest.fn(() => null)
     cache.delete = jest.fn(() => null)
-    models.users = { findByPk: jest.fn(async () => ({ dataValues: { id: 'bar' } })) }
     models.applications = { findByPk: jest.fn(async () => ({ dataValues: { id: 'bar' } })) }
     models.sites = { findByPk: jest.fn(async () => ({ dataValues: { id: 'bar' } })) }
     models.applicationSites = {
       findOne: jest.fn(async () => null),
       create: jest.fn(() => ({
         dataValues: {
-          userId: USER_ID,
           siteId: SITE_ID,
           applicationId: APPLICATION_ID,
           id: APPLICATION_SITE_ID,
@@ -78,11 +75,9 @@ describe('The postApplicationSite handler', () => {
       applicationId: APPLICATION_ID,
       id: expect.any(String),
       siteId: SITE_ID,
-      updateStatus: 'L',
-      userId: USER_ID
+      updateStatus: 'L'
     })
-    expect(cache.save).toHaveBeenCalledWith(`/user/${USER_ID}/application-site/${APPLICATION_SITE_ID}`, expect.any(Object))
-    expect(cache.delete).toHaveBeenCalledWith(`/user/${USER_ID}/application-sites`)
+    expect(cache.save).toHaveBeenCalledWith(`/application-site/${APPLICATION_SITE_ID}`, expect.any(Object))
     expect(h.response).toHaveBeenCalledWith(expect.objectContaining({
       applicationId: APPLICATION_ID,
       id: APPLICATION_SITE_ID,
@@ -93,8 +88,7 @@ describe('The postApplicationSite handler', () => {
     expect(codeFunc).toHaveBeenCalledWith(201)
   })
 
-  it('returns a conflict and error if the user-application-site already exists', async () => {
-    models.users = { findByPk: jest.fn(async () => ({ dataValues: { id: 'bar' } })) }
+  it('returns a conflict and error if the application-site already exists', async () => {
     models.applications = { findByPk: jest.fn(async () => ({ dataValues: { id: 'bar' } })) }
     models.sites = { findByPk: jest.fn(async () => ({ dataValues: { id: 'bar' } })) }
     models.applicationSites = {
@@ -113,7 +107,6 @@ describe('The postApplicationSite handler', () => {
   })
 
   it('returns a bad request and error if the site does not exist', async () => {
-    models.users = { findByPk: jest.fn(async () => ({ dataValues: { id: 'bar' } })) }
     models.applications = { findByPk: jest.fn(async () => ({ dataValues: { id: 'bar' } })) }
     models.sites = { findByPk: jest.fn(async () => null) }
     const postApplicationSite = (await import('../post-application-site.js')).default
@@ -129,7 +122,6 @@ describe('The postApplicationSite handler', () => {
   })
 
   it('returns a bad request and error if the application does not exist', async () => {
-    models.users = { findByPk: jest.fn(async () => ({ dataValues: { id: 'bar' } })) }
     models.applications = { findByPk: jest.fn(async () => null) }
     const postApplicationSite = (await import('../post-application-site.js')).default
     await postApplicationSite(context, req, h)
@@ -143,17 +135,8 @@ describe('The postApplicationSite handler', () => {
     expect(codeFunc).toHaveBeenCalledWith(400)
   })
 
-  it('returns a bad request if the user does not exist', async () => {
-    models.users = { findByPk: jest.fn(async () => null) }
-    const postApplicationSite = (await import('../post-application-site.js')).default
-    await postApplicationSite(context, req, h)
-    expect(h.response).toHaveBeenCalled()
-    expect(typeFunc).toHaveBeenCalledWith(applicationJson)
-    expect(codeFunc).toHaveBeenCalledWith(400)
-  })
-
   it('throws with any model error', async () => {
-    models.users = { findByPk: jest.fn(async () => { throw new Error() }) }
+    models.applications = { findByPk: jest.fn(async () => { throw new Error() }) }
     const postApplicationSite = (await import('../post-application-site.js')).default
     await expect(async () => {
       await postApplicationSite(context, req, h)
