+++ packages/api/src/handlers/application-site/application-site.js
@@ -1,10 +1,10 @@
 import postApplicationSite from './post-application-site.js'
-import getApplicationSitesByUserId from './get-application-sites-by-user-id.js'
+import getApplicationSites from './get-application-sites.js'
 import getApplicationSiteByApplicationSiteId from './get-application-site-by-application-site-id.js'
 import deleteApplicationSiteByApplicationSiteId from './delete-application-site-by-application-site-id.js'
 
 export {
-  getApplicationSitesByUserId,
+  getApplicationSites,
   getApplicationSiteByApplicationSiteId,
   postApplicationSite,
   deleteApplicationSiteByApplicationSiteId
